+++ src/server/pathway-commons/index.js
@@ -1,8 +1,19 @@
 const pc = require('pathway-commons');
 const _ = require('lodash');
+const qs = require('querystring');
+const fetch = require('node-fetch');
 
 const search = require('./search/');
 const datasources = require('./datasources');
+
+
+const fetchOptions = {
+  method: 'GET',
+  headers: {
+    'Content-Type': 'application/json',
+    'Accept': 'application/json'
+  }
+};
 
 const PathwayCommonsService = {
   // query pathway commons for pathways, sbgn, information, etc.
@@ -29,6 +40,39 @@ const PathwayCommonsService = {
   // check if pathway commons is online
   isServiceOnline (delay) {
     return pc.utilities.pcCheck(delay);
+  },
+
+  get(options) {
+    //Construct query url
+    let url = 'https://www.pathwaycommons.org/pc2/get?' + qs.stringify(options);
+    return fetch(url, fetchOptions).then(response => response.text());
+  },
+  
+  //Queries PC2 using traverse
+  //Returns metadata for the pathway
+  //Requires a valid query and uri
+  traverse(options) {
+    //Construct query url
+    let url = 'https://www.pathwaycommons.org/pc2/traverse?' + qs.stringify(options);
+    return fetch(url, fetchOptions).then(response => response.json());
+  },
+  
+  search(options) {
+    //Construct query url
+    let url = 'https://www.pathwaycommons.org/pc2/search?' + qs.stringify(options); 
+    return fetch(url, fetchOptions).then(response => response.json());
+  },
+  
+  graph(options) {
+    //Construct query url
+    let url = 'https://www.pathwaycommons.org/pc2/graph?' + qs.stringify(options); 
+    return fetch(url, fetchOptions).then(response => response.json());
+  },
+  
+  top_pathways(options) {
+    //Construct query url
+    let url = 'https://www.pathwaycommons.org/pc2/top_pathways?' + qs.stringify(options); 
+    return fetch(url, fetchOptions).then(response => response.json());
   }
 };
